api/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "daMpoff",email = "11projectone@gmail.com"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "mistralai (>=0.0.10,<1.0.0)"
]

[tool.poetry]
packages = [{include = "src", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
